Run full Code Cleanup on the whole project
This cleanup is done so we don't have to do it every time we change an
old file.

The only thing that was not included in this Code Cleanup from
"Visual Studio > Configure Code Cleanup" was "Apply namespace matches
folder preferences" because changing the file structure is a dangerous
process.

Testing done: Loaded an old save, checked basic functionality, updated
milestones, saved changes, checked that they load on the previous
version.