refactor: move to first page if no page active
Also, fix a data consistency issue - the page delete button in the page settings pane wasn't removing the pane from the display pages list, just marking it for deletion. The page still wouldn't be displayed but it would be in the list, which wasn't an issue when we just pick items out of the list based on what's displayed but became an issue when we iterate through it.