fix: improve the logic of closing a utility window
On Linux, the file-picker in the Welcome Window would close right
after opening it.

This patch changes the logic of when the utility windows are closed.
Previously, they closed on losing focus. That likely was the reason
for the bug because they lost focus on creation. Now they don't
close on losing focus.

Co-authored-by: Dale McCoy <21223975+DaleStan@users.noreply.github.com>